fixup! ISSUE-1716 OIDC set up for JMAP endpoint
