+++ tmail-backend/apps/distributed/src/main/conf/jmap.properties
@@ -40,7 +40,7 @@ jwt.privatekeypem.url=file://conf/jwt_privatekey
 
 # If you want to specify authentication strategies for Jmap rfc-8621 version
 # For custom Authentication Strategy not inside package "org.apache.james.jmap.http", you have to specify its FQDN
-authentication.strategy.rfc8621=JWTAuthenticationStrategy,BasicAuthenticationStrategy,com.linagora.tmail.james.jmap.ticket.TicketAuthenticationStrategy
+authentication.strategy.rfc8621=JWTAuthenticationStrategy,BasicAuthenticationStrategy,com.linagora.tmail.james.jmap.ticket.TicketAuthenticationStrategy,com.linagora.tmail.james.jmap.oidc.OidcAuthenticationStrategy
 
 # Gives an URL for OpenID discovery being exposed on .well-known/webfinger endpoint
 # CF https://openid.net/specs/openid-connect-discovery-1_0.html
