+++ tmail-backend/integration-tests/jmap/jmap-integration-tests-common/src/main/java/com/linagora/tmail/common/OidcAuthenticationContract.java
@@ -23,9 +23,6 @@ import static io.restassured.RestAssured.requestSpecification;
 import static org.apache.james.jmap.JMAPTestingConstants.BOB;
 import static org.apache.james.jmap.JMAPTestingConstants.BOB_PASSWORD;
 import static org.apache.james.jmap.JMAPTestingConstants.DOMAIN;
-import static org.apache.james.jmap.core.JmapRfc8621Configuration.UPLOAD_LIMIT_DEFAULT;
-import static org.apache.james.jmap.core.JmapRfc8621Configuration.URL_PREFIX_DEFAULT;
-import static org.apache.james.jmap.core.JmapRfc8621Configuration.WEBSOCKET_URL_PREFIX_DEFAULT;
 import static org.apache.james.jmap.rfc8621.contract.Fixture.AUTHORIZATION_HEADER;
 import static org.apache.james.jmap.rfc8621.contract.Fixture.ECHO_REQUEST_OBJECT;
 import static org.apache.james.jmap.rfc8621.contract.Fixture.baseRequestSpecBuilder;
@@ -36,10 +33,8 @@ import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.time.Clock;
 import java.time.Duration;
-import java.util.Optional;
 
 import org.apache.james.GuiceJamesServer;
-import org.apache.james.jmap.core.JmapRfc8621Configuration;
 import org.apache.james.utils.DataProbeImpl;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
