+++ tmail-backend/jmap/extensions/src/main/java/com/linagora/tmail/james/jmap/oidc/TokenInfo.java
@@ -16,7 +16,22 @@
  *  more details.                                                   *
  ********************************************************************/
 
-package com.linagora.tmail.james.jmap.model;
+package com.linagora.tmail.james.jmap.oidc;
 
-public record Token(String value) {
+import java.time.Instant;
+import java.util.List;
+import java.util.Optional;
+
+import com.google.common.base.MoreObjects;
+
+public record TokenInfo(String email, Optional<Sid> sid, Instant exp, List<Aud> aud) {
+
+    public String asString() {
+        return MoreObjects.toStringHelper(this)
+            .add("email", email)
+            .add("sid", Optional.ofNullable(sid).flatMap(sidValue -> sidValue.map(Sid::value)).orElse(null))
+            .add("exp", exp)
+            .add("aud", Optional.ofNullable(aud).orElseGet(List::of).stream().map(Aud::value).toList())
+            .toString();
+    }
 }
